# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SynthChat Hermes Backend API
  version: 0.1.0
  description: |
    Frontend-facing contract for the complete Rust backend. Upstream Hermes
    HTTP, dashboard, WebSocket, CLI, filesystem and SQLite interfaces are
    compatibility references only and are not production runtime dependencies.
servers:
  - url: http://127.0.0.1:8642
    description: Local desktop backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/web/providers:
    get:
      tags: [Web]
      operationId: listWebProviders
      summary: List Rust web provider adapters
      responses:
        "200":
          description: Implemented Rust web provider catalog without credentials. The Web-first contract exposes Tavily only.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WebProvider" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/profiles/{profileId}/web:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Web]
      operationId: getProfileWebConfig
      summary: Get Profile web provider configuration and readiness
      responses:
        "200":
          description: Non-secret Web configuration and effective per-tool readiness. ETag is the shared ProfileConfig revision.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    patch:
      tags: [Web]
      operationId: updateProfileWebConfig
      summary: Update Profile web provider configuration
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/WebConfigPatch" }
      responses:
        "200":
          description: Updated non-secret Web configuration. Unknown YAML fields are preserved and ETag is the resulting shared ProfileConfig revision.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebConfig" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/v1/profiles/{profileId}/toolsets:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Tools]
      operationId: listToolsets
      summary: List dynamic toolsets
      responses:
        "200":
          description: Dynamic toolset catalog for a profile. ETag is the current ProfileConfig revision shared with config and toolset writes.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Toolset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/profiles/{profileId}/toolsets/{toolsetId}:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
      - $ref: "#/components/parameters/ToolsetId"
    patch:
      tags: [Tools]
      operationId: updateToolset
      summary: Update toolset enablement
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/ToolsetPatch" }
      responses:
        "200":
          description: Updated toolset. ETag is the resulting ProfileConfig revision; an enablement no-op returns the unchanged revision.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Toolset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
  /api/v1/profiles/{profileId}/skills:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Skills]
      operationId: listSkills
      summary: List installed skills
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 500 } }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Installed skills for a profile. ETag combines the current ProfileConfig revision with the complete unfiltered installed-skill catalog and is used by Skill PATCH.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/profiles/{profileId}/skills/install:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    post:
      tags: [Skills]
      operationId: installSkill
      summary: Start a skill installation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallSkill" }
      responses:
        "202":
          description: Installation operation started, not completed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/profiles/{profileId}/skills/{skillId}:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
      - $ref: "#/components/parameters/SkillId"
    patch:
      tags: [Skills]
      operationId: updateSkill
      summary: Update a skill
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/SkillPatch" }
      responses:
        "200":
          description: Updated skill state/configuration. ETag is the resulting ProfileConfig revision.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
    delete:
      tags: [Skills]
      operationId: uninstallSkill
      summary: Start a skill uninstall
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Uninstall operation started.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-desktop-session-token
  headers:
    ETag:
      description: Opaque resource revision, including quotes as required by HTTP.
      schema: { type: string }
  parameters:
    ProfileId: { name: profileId, in: path, required: true, schema: { $ref: "#/components/schemas/ProfileIdValue" } }
...
    ToolsetId: { name: toolsetId, in: path, required: true, schema: { type: string } }
...
    SkillId: { name: skillId, in: path, required: true, schema: { type: string, pattern: '^skill_[0-9a-f]{32}$' } }
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque, tamper-evident continuation bound to the endpoint filters and snapshot boundary.
      schema: { type: string, minLength: 1, maxLength: 4096 }
    Limit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: A single quoted strong ETag. Weak, wildcard, multiple or malformed values are invalid.
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for non-idempotent creates. Scoped by installation, method and canonical path; retained for at least 24 hours. Reuse with a different request fingerprint returns 409.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    BadRequest:
      description: Invalid request.
      x-error-codes: [validation_failed, invalid_if_match, invalid_cursor]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing or invalid desktop token.
      x-error-codes: [unauthorized]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      x-error-codes: [resource_not_found, profile_not_found, approval_not_found, clarification_not_found, hermes_state_not_found]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: State, revision, idempotency, replay-window or busy conflict.
      x-error-codes: [revision_conflict, idempotency_conflict, idempotency_resource_gone, mcp_server_exists, mcp_config_invalid, capacity_exceeded, unsafe_file_path, memory_storage_drift, profile_delete_conflict, session_busy, session_archived, event_history_expired, hermes_import_source_changed, hermes_import_conflict]
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PreconditionRequired:
      description: A single strong If-Match value is required.
      x-error-codes: [precondition_required]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ServiceUnavailable:
      description: A required local service or protected store is unavailable.
      x-error-codes: [secret_storage_unavailable, file_storage_unavailable, skill_storage_unavailable, skill_management_unavailable, operation_storage_unavailable]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Unprocessable:
      description: A required capability/provider is unavailable, or a semantically valid request violates a protected resource constraint.
      x-error-codes: [engine_capability_missing, memory_provider_unsupported, memory_content_blocked, memory_capacity_exceeded, session_search_unavailable, hermes_schema_unsupported, hermes_import_source_invalid, hermes_attachments_require_policy]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    TooManyRequests:
      description: Run, queue, or bounded asynchronous operation capacity exceeded.
      x-error-codes: [capacity_exceeded, skill_source_rate_limited, skill_operation_capacity]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    WebProviderId:
      type: string
      const: tavily
      description: Provider ID accepted by WebConfigPatch in the Web-first slice.
    WebProvider:
      type: object
      additionalProperties: false
      required: [id, displayName, supportsSearch, supportsExtract, secretNames, defaultBaseUrl, customEndpointSupported]
      properties:
        id: { $ref: "#/components/schemas/WebProviderId" }
        displayName: { type: string, const: Tavily }
        supportsSearch: { type: boolean, const: true }
        supportsExtract: { type: boolean, const: true }
        secretNames:
          type: array
          minItems: 1
          maxItems: 1
          items: { type: string, const: TAVILY_API_KEY }
        defaultBaseUrl:
          type: string
          format: uri
          maxLength: 2048
          pattern: "^[Hh][Tt][Tt][Pp][Ss]://"
          description: Backend-effective Tavily base URL. It defaults to the official endpoint and may be set only by trusted runtime configuration; it is not Profile-writable.
        customEndpointSupported:
          type: boolean
          const: false
          description: Profiles cannot supply or override provider endpoints; trusted runtime configuration is outside the Profile API.
    EffectiveWebProvider:
      type: object
      additionalProperties: false
      required: [providerId, status, missingSecretNames]
      properties:
        providerId:
          type: [string, "null"]
          maxLength: 128
          description: Effective provider ID after capability-specific override, shared fallback and the single implemented-provider shortcut. An unsupported legacy YAML value is preserved here for diagnostics.
        status:
          type: string
          enum: [ready, unconfigured, missingSecret, unsupported, capabilityUnsupported]
          description: Readiness is a local configuration/keychain projection and performs no billable or remote network probe.
        missingSecretNames:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/SecretName" }
          description: Required keychain names that are absent. Values, previews and lengths are never returned.
    WebConfig:
      type: object
      additionalProperties: false
      required: [revision, sharedProvider, searchProvider, extractProvider, extractCharLimit, effectiveSearch, effectiveExtract]
      properties:
        revision:
          type: string
          minLength: 1
          description: Unquoted shared ProfileConfig revision corresponding to the response ETag.
        sharedProvider:
          type: [string, "null"]
          maxLength: 128
          description: Stored web.backend fallback. Null means no explicit shared provider; unsupported legacy values remain visible but are never executed.
        searchProvider:
          type: [string, "null"]
          maxLength: 128
          description: Stored web.search_backend override. Null delegates to sharedProvider and then the single implemented-provider shortcut.
        extractProvider:
          type: [string, "null"]
          maxLength: 128
          description: Stored web.extract_backend override. Null delegates to sharedProvider and then the single implemented-provider shortcut.
        extractCharLimit:
          type: integer
          minimum: 2000
          maximum: 500000
          default: 15000
          description: Per-page clean character budget for web_extract provider content.
        effectiveSearch: { $ref: "#/components/schemas/EffectiveWebProvider" }
        effectiveExtract: { $ref: "#/components/schemas/EffectiveWebProvider" }
    WebConfigPatch:
      type: object
      additionalProperties: false
      description: Dedicated RFC 7396 merge patch for non-secret Web settings. Null clears a provider override. The only selectable provider is tavily; existing unsupported YAML values are preserved until explicitly replaced or cleared.
      x-merge-patch-profile:
        explicitNullPaths: [/sharedProvider, /searchProvider, /extractProvider]
        nullDeletesProperties: true
        rejectOtherNulls: true
      properties:
        sharedProvider:
          oneOf:
            - $ref: "#/components/schemas/WebProviderId"
            - type: "null"
        searchProvider:
          oneOf:
            - $ref: "#/components/schemas/WebProviderId"
            - type: "null"
        extractProvider:
          oneOf:
            - $ref: "#/components/schemas/WebProviderId"
            - type: "null"
        extractCharLimit:
          type: integer
          minimum: 2000
          maximum: 500000
...
    ProfileIdValue:
      type: string
      pattern: "^(default|[a-z0-9_][a-z0-9_-]{0,63})$"
...
    SecretName:
      type: string
      pattern: "^[A-Z][A-Z0-9_]{0,127}$"
...
    Toolset:
      type: object
      additionalProperties: false
      required: [id, displayName, description, enabled, configured, tools]
      properties:
        id: { type: string }
        displayName: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        configured:
          type: boolean
          description: >-
            Whether this installation currently satisfies the Toolset's local
            configuration prerequisite. For code_execution this means a usable
            host Python 3 interpreter was detected; callers must still require
            capabilities.extensions.codeExecution=true and enabled=true before
            presenting execute_code as runnable.
        tools: { type: array, items: { type: string } }
        configSchema: { type: object, additionalProperties: true }
    ToolsetPatch:
      type: object
      additionalProperties: false
      required: [enabled]
      description: Exact enablement update. General Toolset configuration is unavailable until a bounded, secret-safe configuration model is defined.
      properties:
        enabled: { type: boolean }
    Skill:
      type: object
      additionalProperties: false
      required: [id, name, description, source, enabled, uninstallable, configurable]
      properties:
        id: { type: string, pattern: '^skill_[0-9a-f]{32}$' }
        name: { type: string }
        description: { type: string }
        source: { type: string, enum: [bundled, file, local, registry, url] }
        version: { type: [string, "null"] }
        enabled: { type: boolean }
        uninstallable: { type: boolean }
        configurable: { type: boolean }
        configSchema: { type: object, additionalProperties: true }
    SkillPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Skill" } }
        nextCursor: { type: [string, "null"] }
    InstallSkill:
      type: object
      additionalProperties: false
      properties:
        registryId: { type: string, minLength: 1, maxLength: 512 }
        url:
          type: string
          format: uri
          pattern: '^https://.*/SKILL\.md$'
          minLength: 1
          maxLength: 2048
          description: Direct public HTTPS SKILL.md URL. Userinfo, percent escapes, query, fragment, redirects to unsafe hosts, and every private or special resolved address are rejected.
        fileId: { type: string, pattern: '^file_[0-9a-f]{32}$' }
      oneOf:
        - required: [registryId]
        - required: [url]
        - required: [fileId]
    SkillPatch:
      type: object
      additionalProperties: false
      properties:
        enabled: { type: boolean }
        config: { type: object, additionalProperties: true }
    Operation:
      type: object
      additionalProperties: false
      required: [id, kind, status, createdAt, updatedAt]
      properties:
        id: { type: string, pattern: '^op_[0-9a-f]{32}$' }
        kind: { type: string, enum: [skillInstall, skillUninstall] }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        error: { oneOf: [{ $ref: "#/components/schemas/Problem" }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId, retryable]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, "null"] }
        instance: { type: [string, "null"], format: uri-reference }
        code:
          type: string
          description: Stable machine-readable code. Each response component's x-error-codes lists known values without closing the set to future additions.
        requestId: { type: string }
        retryable: { type: boolean }